# entrega-04/api-server/std/api.yaml
openapi: 3.0.1
info:
  title: Projeto-Final-STD
  description: |
    API REST para gerenciar dispositivos IoT e grupos de associação. 
    Oferece operações CRUD para dispositivos individuais (IOTs) e grupos, 
    permitindo a criação, atualização e exclusão de recursos.
  version: v0
servers:
- url: http://api-server:8080
  description: IOT-Controller-STD
paths:
  /iot:
    get:
      tags:
      - iot-controller
      summary: Retorna todos os IOTs
      description: Retorna todos os iots presentes na api
      operationId: obterTodosIot
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/prodIOT'
    put:
      tags:
      - iot-controller
      summary: Atualiza um IOT existente
      description: É possivel alterar os dados de um iot existente na api
      operationId: atualizarStatusProdIOT
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/prodIOT'
    post:
      tags:
      - iot-controller
      summary: Cria um produto IOT
      description: Cria um produto IOT
      operationId: adicionarIot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/prodIOT'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/prodIOT'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/prodIOT'
  /iot/{id}:
    get:
      tags:
      - iot-controller
      summary: Retorna um determinado IOT
      description: Retorna um determinado IOT
      operationId: obterIot
      parameters:
      - name: id
        in: path
        description: id do iot
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/prodIOT'
    delete:
      tags:
      - iot-controller
      summary: Remove um IOT
      description: Remove um IOT
      operationId: excluirIot
      parameters:
      - name: id
        in: path
        description: id do iot
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: No Content
  /grupo:
    get:
      tags:
      - iot-controller
      summary: Retorna todos os grupos criados
      description: Retorna todos os grupos criados
      operationId: obterTodosGrupo
      parameters: []
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupIOT'
    post:
      tags:
      - iot-controller
      summary: Cria um grupo de associação
      description: Cria um grupo de associação
      operationId: criaGrupo
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupIOT'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GroupIOT'
    put:
      tags:
      - iot-controller
      summary: Atualiza dados do grupo 
      description: Pode atualizar o grupo e adicionar produtos
      operationId: atualizarStatusDeTodosIotsDoGrupo
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupIOT'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupIOT'
  /grupo/{grupoId}:
    get:
      tags:
      - iot-controller
      summary: Retorna um determinado grupo 
      description: Retorna um determinado grupo 
      operationId: obterTodosIotNoGrupo
      parameters:
      - name: grupoId
        in: path
        description: id do grupo
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupIOT'

    delete:
      tags:
      - iot-controller
      summary: Deleta um grupo 
      description: Remove um grupo sem afetar os produtos
      operationId: excluirGrupo
      parameters:
      - name: id
        in: path
        description: id do grupo
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: No Content
components:
  schemas:
    prodIOT:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        status:
          type: string
        dadosAtual:
          type: integer
          format: int32
        metodos:
          type: array
          items: 
            type: string
    GroupIOT:
      type: object
      properties:
        grupoId:
          type: integer
          format: int64
        nome:
          type: string
        produtos:
          type: array
          items:
            $ref: '#/components/schemas/prodIOT'